+++ controllers/ProfessorController.php
@@ -64,6 +64,10 @@ class ProfessorController extends Controller
      */
     public function actionCreate()
     {
+        if(Yii::$app->user->identity->username != 'admin'){
+            throw new NotFoundHttpException('Você não tem permissão para acessar esta página!');
+        }
+
         $model = new Professor();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -84,6 +88,10 @@ class ProfessorController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(Yii::$app->user->identity->username != 'admin'){
+            throw new NotFoundHttpException('Você não tem permissão para acessar esta página!');
+        }
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -104,6 +112,10 @@ class ProfessorController extends Controller
      */
     public function actionDelete($id)
     {
+        if(Yii::$app->user->identity->username != 'admin'){
+            throw new NotFoundHttpException('Você não tem permissão para acessar esta página!');
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
